Update heatMon.py
There are some things missing for a "publicly available code". However, I choose to publish it anyway.

Things that needs to be added:

MQTT Security
Use of static variables for setting USB-port, MQTT user/pass etc.

But for the time being, you have to find it in the beginning of the code and add it.